Make sure logging & sentry initalization is correct and consistent
