+++ server/bleep/Cargo.toml
@@ -14,6 +14,7 @@ ee-cloud = ["ee-pro"]
 
 [[bin]]
 name = "bleep"
+required-features = ["color-eyre"]
 
 [[bench]]
 name = "snippets"
@@ -43,7 +44,7 @@ clap = { version = "4.4.4", features = ["derive"] }
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.17", features = ["env-filter", "registry"] }
 tracing-appender = "0.2.2"
-color-eyre = "0.6.2"
+color-eyre = { version = "0.6.2", optional = true }
 regex = "1.9.5"
 regex-syntax = "0.6.29"
 fuzzy-matcher = "0.3.7"
