+++ server/bleep/src/bin/bleep.rs
@@ -5,6 +5,8 @@ use bleep::{Application, Configuration, Environment};
 async fn main() -> Result<()> {
     let config = Configuration::cli_overriding_config_file()?;
 
+    _ = color_eyre::install();
+
     Application::install_logging(&config);
     let app = Application::initialize(Environment::server(), config, None, None).await?;
